feat: introduce inter-process locking in coreclient (#779)
The outbound service now runs exclusively across multiple processes. In
particular, no races can occur when multiple instances of the app are running
(Desktop) or when an additional background extension is active (Mobile).

The locking mechanism is based on a file in the client's data directory. On
iOS, it is important that this file is shared between the app and the
extension.