perf: load chats details from an LRU cache (#775)
This commit introduces a chats repository -- a simple LRU cache used to
create a `ChatDetailsCubit` synchronously with initial data if the
corresponding chat details are already in the repository. This
eliminates flickering when switching between chats.

The chat list now loads only an ordered list of chat IDs. Each list tile
then creates a `ChatDetailsCubit` from the chat ID and the chats
repository.

The chats repository API is not exposed to Dart and is only used inside
Rust. However, the lifetime of the repository is bound to the widget
tree. It is mounted near the root and available throughout the app. Chat
data in the repository is updated whenever a `ChatDetailsCubit` updates
its own data.

Also fix a comparison implementation of `ImageData`. The field `data`
was compared by identity instead as bytes. Now, we only compare the
`hash`.