fix: out of order chat list (#780)
This commit fixes two issues:

A draft message could have an `updated_at` value even when it was empty
(i.e., no text and no edits). This check was missing, which caused chats
that had a draft message *at some point in time* to be pushed to the top
of the list.

The second issue was that chat list tiles contained stale data whenever
the list was reordered. This was caused by the cached chat details cubit
in the list tile.