fix: chat list reordering when typing the message
Previously, while typing a message, we stored the message draft in the
database. This triggered a reordering of the chat list, since chats
containing a draft were shown at the top. This behavior was confusing
because there was no visible change in the chat list, yet it still
reordered.

Now, we store the draft but mark it as *committed* only when leaving the
chat. Non-committed drafts are not considered when ordering the chat
list, and no chat store updates are sent for them. Non-committed drafts
are also committed on app startup to ensure consistency if the app is
closed while composing a message.