refactor: remove grpc port and add `listen` server config
Server is now configured with a `listen` parameter, which is a socket
address — that is, an IP address and a port. The default listen address
is `127.0.0.1:8080` for local development and `0.0.0.0:80` for
production.

The Kubernetes deployment has been updated to redirect port 443 to 80
instead of using a custom gRPC port. HTTP traffic has been removed.