feat: add Prometheus metrics
This commit introduces Prometheus metrics for the gRPC server. The
metrics are exposed on a separate TCP address under the `/metrics`
route.

It also cleans up the logging setup.

Implementation details:

Metrics are collected using the `metrics` crate in an additional tonic
layer. The `GrpcMetricsLayer` records metrics such as RPC start,
completion, and handling latency. The Prometheus exporter is integrated
using Axum to expose metrics on a dedicated address (default:
`127.0.0.1:9090`). The exporter service provided by
`metrics-exporter-prometheus` is not used to avoid introducing an
additional HTTP/TLS stack.