feat(server): add max attachment size limit (#811)
The server storage configuration includes a new option,
`max_attachment_size`, which limits the maximum size of an attachment in
bytes. The default value is 20 MiB.

The size of an attachment is enforced when uploading via a POST request
containing a signed policy. The previous method of uploading attachments
via a PUT request is still supported but will be removed in the future.