hotfix(server): stale qs listeners are not cleaned up
This is a regression introduced in #808. Previously, we stopped tracking
a queue listener when we got an error in the incoming requests handler
in QS. When error handling was suppressed, we stopped calling `untrack
listener`, which is the bug. As a consequence, no push notifications
were sent to the client because it was considered as connected to the
server.

However, it is better to tie the listener's liveness to the lifetime of
the queue context, which is owned by the queue listen stream. Now, when
gRPC drops the stream, the context is dropped and cancels the listener.
It also unsubscribes from the Postgres notifications.

Also, debug logs have been added to the queue listener and PG listener
task.