fix(app): fix mark as read datetime truncation corner case (#818)
Timestamps of messages are sent from Rust to Dart and then back to Rust
for marking a chat as read. Since Dart only supports microsecond
precision, the timestamp might be truncated, which can lead to a
different ordering of two messages. We handle this corner case, even
though it is unlikely to happen.