feat: save attachments to file
Added a context menu for file and image attachment to save them to the
device. On Android, attachments are saved in the "Air" folder the
Downloads/Pictures directory. On iOS, the attachments are saved in the
Document directory of the application container. On Desktop, a dialog a
show to select a directory to save the attachments.