feat(coreclient): connection requests via targeted messages (#846)
Allow clients to establish a connection with another client that they're
in a group with. For now, a **targeted message** is an application
message (and thus decryptable by anyone in the given group) fanned out
to only a single group member.

- Add a new endpoint to the DS that allows clients to send targeted
messages
- Add new functions to the `Store` that allow adding a new contact via a
targeted message and listing pending connection requests made through
targeted messages
- Add the machinery necessary to compose and process targeted messages