feat(app): save attachment context menu (#844)
Added a context menu for file and image attachments to save them to the
device. On Android, attachments are saved in the "Air" folder within the
Downloads/Pictures directory. On Desktop, a dialog is shown to select a
directory to save the attachments. On iOS, saving of attachments is not
supported.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>