feat(app): improve saving of attachments on Android (#864)
We no longer just write to a file in Rust; instead, we use the
Android-specific ContentResolver API to acquire a URI and write to it.
This change immediately makes the media available in Files and other
apps. Furthermore, it automatically manages duplicate filenames and
determines the correct path based on the MIME type.