feat(backend): introduce client metadata in RPCs (#878)
Client metadata contains the client version, platform, and release
channel. It
is attached to every gRPC request and can be used to identify the client
and
possibly deny serving it. A typical size in bytes of the client metadata
is about
16 bytes.

Also remove protobuf messages unused in RPCs. Consolidate the apiclient
modules
and reduce the `ApiClient` size on the stack.

Warning: a newer client cannot be used with an older server. The older
server
won't be able to verify the signature without client metadata.