fix(app): various papercuts (#874)
* Fix keyboard content overflow on Android
* Fix the app bar height in the chat list view and message list.
* Fix the dark mode chat view test.
* Fix conflict between scroll gesture and app overview gesture on
Android

Also fix design on the intro, sign up, and username screens:

* Add colors to text in buttons and progress indicator
* Fix size of text in buttons and progress indicator
* Align bottom padding of buttons
* Always show avatar picker

Note about bottom screen padding: On screens that resize when showing
the on-screen keyboard and push a button or other components upward,
it’s not ideal to use dynamic padding based on keyboard visibility. This
creates a choppy animation. Instead, it’s better to use static padding
at the bottom below the last component while also using a safe area.
When the keyboard is not shown, the safe area adds padding on top of the
static padding; when the keyboard is shown, only the static padding
remains.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>