feat(app): add chat button to contact details (#860)
The button opens 1:1 chat with the user, if the user is a contact.
Otherwise, a dialog is shown to send a contact request. After contact
request is sent, the newly created chat is opened.

Also add handling of targeted message connection chats.

Additionaly, fix the design of the error banner:

* Remove bottom separator and elevation
* Set color of the error message

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>