feat(app): connection requests (#898)
This commit introduces a dialog for accepting pending connection
requests. To support this, we introduced a new chat type
`PendingConnection`, which represents an incoming connection request,
either via a user name or a targeted message in another chat. Such a
chat does not have an underlying group (the group ID of the chat is set
to the incoming group ID, but the group is not created yet). All
information needed for accepting the connection request is stored in the
client database (new table `pending_connection_info`).

The system messages `ReceivedDirectConnectionRequest` and
`ReceivedHandleConnectionRequest` are rendered as a connection request
dialog for chats of type `PendingConnection`. When the user taps
`Accept`, the pending connection information is loaded, and a new MLS
group is created by joining it via an external commit. The commit is
sent to DS, which communicates to the connection sender that the user
has accepted the request.

Also fix a bug where the chat title for received connection request
system messages was taken from both the new chat and the original chat.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>