hotfix: new users have invalid domain
The bug was introduced in #800 where we started redirecting from air.ms
to prod.air.ms in the registration cubit. This is however wrong, since
then new users have prod.air.ms as domain, which is not valid.

This fix removes the faulty redirection and does it only in the check
validation code function.