feat(app): sort user lists in alphabetical order (#889)
Sorts users in lists by display name in alphabetical order. Note that
this is done in dart, because we don't have have aggregated data of user
IDs and their corresponding display names at the DB/Rust level and it
would be more expensive in terms of IO ops to do it there.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: boxdot <dimitri@phnx.im>