fix(server): race between listen and unlisten PG in QS
The was a race between listen und unlisten Postgres when evicting
another listening client in QS. In case, the order of execution was
listen and then unlisten, the new connection was not notified about the
new messages in the queue. As consequence, the new connection was not
receiving any messages. This race was usually triggered when the client
changed its network (e.g. connected to a different network via VPN).