fix: race condition in QS postres listen/unlisten (#919)
There was a race between LISTEN and UNLISTEN in Postgres when evicting a
listening client in QS. In cases where the order of execution was LISTEN
and then UNLISTEN, the new connection was not notified about new
messages in the queue. As a consequence, the new connection did not
receive any messages. This race was usually triggered when the client
changed its network (e.g., connecting to a different network via VPN).