feat(app): new icon system (#912)
This PR domesticates the icons used in the app by exposing them through
an `AppIcon` widget.
It replaces the `iconoir` icon set through customized icons, drawn from
the Lucide icon set.
To be future-proof and support custom-drawn icons, all icons are
provided in the SVG format. This PR includes a script to generate a more
efficient binary representation and collect the icons into the
`AppIconType` enum.
The icons can be generated through the new `just generate-icons`
command.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>