feat(app): drop read receipts (#931)
Continues the implementation of the read receipt setting (follow-up of
#794). Read receipts are now dropped upon reception when the setting is
off. There is a test that checks that changing the value of the setting
has an effect.

The PR pushes `ReadReceiptsSetting` into the core client. This seemed
like the least intrusive way to implement this, but there are definitely
alternatives.